added post api tests for statuses + fixed import issues for api tests
