+++ tests/ui/test_get_public_gists.py
@@ -5,13 +5,14 @@ import logging
 
 from playwright.sync_api import sync_playwright, expect
 
-from helpers.bodies import PRIVATE_GIST
-from helpers.create_gist import create_gist
-from helpers.delete_gist import delete_gist
+from helpers_api.bodies import PUBLIC_GIST
+from helpers_api.create_gist import create_gist
+from helpers_api.delete_gist import delete_gist
 import allure
 
-from helpers_ui.models.login_page import LoginPage
+from helpers.models.login_page import LoginPage
 from logs.logging_config import configure_logging
+from tests.ui.models.dicover_gists_page import DiscoverGistsPage
 from tests.ui.models.gist_username_page import GistUserNamePage
 
 configure_logging()
@@ -29,7 +30,7 @@ BROWSER_TYPES = ["chromium", "webkit"]
 @allure.step("Create gist")
 @pytest.fixture(scope="session")
 def setup_gist():
-    body = PRIVATE_GIST
+    body = PUBLIC_GIST
     # Create a gist
     gist_id = create_gist(body, GITHUB_TOKEN)
     os.environ['GIST_ID'] = gist_id
@@ -46,13 +47,30 @@ def teardown_gist():
         delete_gist(gist_id, GITHUB_TOKEN)
 
 
-@allure.feature("/GET Private Gists")
-class TestPrivateGists:
+@allure.feature("/GET Public Gists")
+class TestPublicGists:
 
-    @allure.story("UI Test: check visibility of a private gist on https://gist.github.com/username for authed user")
-    @allure.severity(allure.severity_level.BLOCKER)
+    @allure.story("UI Test: check visibility of a public gist on https://gist.github.com/discover for NOT authed user")
+    @allure.severity(allure.severity_level.NORMAL)
     @pytest.mark.parametrize("browser_type", BROWSER_TYPES)
-    def test_get_private_gist_on_username_page_by_authed_user(self, browser_type, setup_gist, teardown_gist):
+    def test_public_gist_visibility_on_discover(self, browser_type, setup_gist, teardown_gist):
+        with sync_playwright() as p:
+            gist_id = setup_gist
+
+            browser = p[browser_type].launch(headless=True)
+            page = browser.new_page()
+
+            discover_gists_page = DiscoverGistsPage(page)
+            discover_gists_page.navigate()
+
+            with allure.step("Check that public gist is visible for stranger"):
+                expect(page.get_by_role("link", name=f"gist:{gist_id}")).to_be_visible()
+            browser.close()
+
+    @allure.story("UI Test: check visibility of a public gist on https://gist.github.com/username for authed user")
+    @allure.severity(allure.severity_level.CRITICAL)
+    @pytest.mark.parametrize("browser_type", BROWSER_TYPES)
+    def test_get_public_gist_on_username_page_by_authed_user(self, browser_type, setup_gist, teardown_gist):
         with sync_playwright() as p:
             gist_id = setup_gist
 
@@ -67,14 +85,14 @@ class TestPrivateGists:
             gist_username_page = GistUserNamePage(page, gist_id, USERNAME_EREKA)
             gist_username_page.navigate(USERNAME_EREKA)
 
-            with allure.step("Check that private gist is visible for creator"):
+            with allure.step("Check that public gist is visible for creator"):
                 expect(page.get_by_role("link", name=f"gist:{gist_id}")).to_be_visible()
             browser.close()
 
-    @allure.story("UI Test: check visibility of a private gist on https://gist.github.com/username for NOT authed user")
-    @allure.severity(allure.severity_level.CRITICAL)
+    @allure.story("UI Test: check visibility of a public gist on https://gist.github.com/username for NOT authed user")
+    @allure.severity(allure.severity_level.BLOCKER)
     @pytest.mark.parametrize("browser_type", BROWSER_TYPES)
-    def test_get_private_gist_on_username_page(self, browser_type, setup_gist, teardown_gist):
+    def test_get_public_gist_on_username_page(self, browser_type, setup_gist, teardown_gist):
         with sync_playwright() as p:
             gist_id = setup_gist
 
@@ -84,6 +102,6 @@ class TestPrivateGists:
             gist_username_page = GistUserNamePage(page, gist_id, USERNAME_EREKA)
             gist_username_page.navigate(USERNAME_EREKA)
 
-            with allure.step("Check that private gist is NOT visible for stranger"):
-                expect(page.get_by_role("link", name=f"gist:{gist_id}")).not_to_be_visible()
+            with allure.step("Check that public gist is visible for stranger"):
+                expect(page.get_by_role("link", name=f"gist:{gist_id}")).to_be_visible()
             browser.close()
